Move FieldIdx and Layout to rustc_target
